Modified SampleScene, added player movement
Added Script for movement based on input system, PlayerController Script to arrange the controls and set up the collision detection that checks whether player is on the ground.
For now a simple cube serves as ground, but it can be modified in any way necessary, just by adding "Ground " Layer to a game object.